Re-arranging orbit counter modules.
