+++ vires/vires/orbit_counter/reader.py
@@ -1,6 +1,6 @@
 #-------------------------------------------------------------------------------
 #
-# Orbit number file handling.
+# Swarm orbit number file handling.
 #
 # Authors: Martin Paces <martin.paces@eox.at>
 #-------------------------------------------------------------------------------
@@ -25,13 +25,9 @@
 # THE SOFTWARE.
 #-------------------------------------------------------------------------------
 
-from os.path import exists, basename, splitext
-from numpy import loadtxt
-from .cdf_util import cdf_open
-from .time_util import mjd2000_to_datetime
-from .aux_common import (
-    SingleSourceMixIn, MJD2000TimeMixIn, BaseReader, render_filename,
-)
+from os.path import exists
+from ..cdf_util import cdf_open
+from ..aux_common import SingleSourceMixIn, MJD2000TimeMixIn, BaseReader
 
 
 class OrbitCounterReader(SingleSourceMixIn, MJD2000TimeMixIn, BaseReader):
@@ -41,53 +37,6 @@ class OrbitCounterReader(SingleSourceMixIn, MJD2000TimeMixIn, BaseReader):
     TYPES = {"orbit": "int32", "Source": "int8"}
     NODATA = {"orbit": -1, "Source": -1}
     INTERPOLATION_KIND = "zero"
-
-
-def update_orbit_counter_file(src_file, dst_file):
-    """ Update Swarm orbit counter text file. """
-    def _write_orbit_counter_file(file_in, src_file, dst_file):
-        with cdf_open(dst_file, "w") as cdf:
-            orbit, time, phi_an, source = parse_orbit_counter_file(file_in)
-            cdf["orbit"], cdf["MJD2000"], cdf["phi_AN"], cdf["Source"] = (
-                orbit, time, phi_an, source
-            )
-            start, end = time.min(), time.max()
-            if src_file:
-                cdf.attrs['SOURCE'] = splitext(basename(src_file))[0]
-            else:
-                cdf.attrs['SOURCE'] = src_file if src_file else render_filename(
-                    "SW_OPER_AUXxORBCNT_{start}_{end}_0001",
-                    mjd2000_to_datetime(start), mjd2000_to_datetime(end)
-                )
-            cdf.attrs['VALIDITY'] = [start, end]
-
-    if isinstance(src_file, str):
-        with open(src_file) as file_in:
-            _write_orbit_counter_file(file_in, src_file, dst_file)
-    else:
-        _write_orbit_counter_file(src_file, None, dst_file)
-
-
-def parse_orbit_counter_file(src_file):
-    """ Parse Swarm orbit counter text file. """
-    data = loadtxt(
-        src_file,
-        skiprows=1,
-        converters={2: (lambda v: 0), 3: (lambda v: 0)}
-    )
-    return (
-        data[:, 0].astype('uint32'), data[:, 1], data[:, 4],
-        data[:, 5].astype('uint8')
-    )
-
-
-def get_max_orbit_number(filename):
-    """ Get maximum orbit number. """
-    if not exists(filename):
-        raise IOError("File %s does not exist!" % filename)
-
-    with cdf_open(filename) as cdf:
-        return cdf["orbit"][-1] - 1
 
 
 def get_orbit_timerange(filename, start_orbit, end_orbit):
@@ -109,3 +58,12 @@ def get_orbit_timerange(filename, start_orbit, end_orbit):
         end_orbit, end_time = _get_ascending_node_time(end_orbit + 1)
         end_orbit -= 1
         return start_orbit, end_orbit, start_time, end_time
+
+
+def get_max_orbit_number(filename):
+    """ Get maximum orbit number. """
+    if not exists(filename):
+        raise IOError("File %s does not exist!" % filename)
+
+    with cdf_open(filename) as cdf:
+        return cdf["orbit"][-1] - 1
